backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pandas==2.2.3
scikit-learn==1.6.1
openai==1.59.3
anthropic==0.42.0
pydantic==2.10.4
python-dotenv==1.0.1
